add db json fetch to cards
